+++ libopenage/util/fds.cpp
@@ -8,7 +8,7 @@
 #include <string.h>
 
 #include <fcntl.h>
-#include "../crossplatform/pty.h"
+#include "pty.h"
 #include <unistd.h>
 
 #include "unicode.h"
